pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimiciii-db"
version = "0.1.0"
description = "DSC 180A Team 1 Paper Reproduction"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Guoxuan Xu", email = "g7xu@ucsd.edu"}
]
dependencies = [
    "sqlalchemy>=2.0.43,<3",
    "psycopg>=3.2.9,<4",
    "psycopg2-binary>=2.9.9,<3",
    "python-dotenv>=1.1.1,<2",
    "jupyterlab>=4.4.9,<5",
    "ipykernel>=6.30.1,<7",
    "pandas>=2.3.3,<3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["mimiciii_db*"]

# Pixi-specific configuration
[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["osx-64", "osx-arm64", "linux-64", "linux-aarch64", "win-64"]

[tool.pixi.dependencies]
python = "3.11.*"
sqlalchemy = ">=2.0.43,<3"
psycopg = ">=3.2.9,<4"
psycopg2-binary = ">=2.9.10,<3"
python-dotenv = ">=1.1.1,<2"
jupyterlab = ">=4.4.9,<5"
ipykernel = ">=6.30.1,<7"
pandas = ">=2.3.3,<3"
matplotlib = ">=3.10.6,<4"
seaborn = ">=0.13.2,<0.14"

[tool.pixi.pypi-dependencies]
mimiciii-db = { path = ".", editable = true }
pytest = "*"
ruff = "*"
black = { extras = ["jupyter"] }
pytest-cov = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.pixi.tasks]
test = "pytest -v"
test-cov = "pytest --cov=mimiciii_db --cov-report=term-missing"
lint = "ruff check ."
fmt = "black ."
